feat: Apply patch from Yanis
